Moved Embedding config to common folder - shared by Quartus and Vivado
